fix(reset-password): fix forgot password
